fix: 병합 과정에서 생기는 문제 해결
