--- app/src/main/java/com/stopsmoke/kekkek/MainActivity.kt
@@ -1,54 +1,0 @@
-package com.stopsmoke.kekkek
-
-import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import androidx.fragment.app.Fragment
-import com.stopsmoke.kekkek.databinding.ActivityMainBinding
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityMainBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
-            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
-            insets
-        }
-
-        setupBottomNavigation()
-
-        if(savedInstanceState == null) {
-            binding.bottomNavigationViewHome.selectedItemId = R.id.home
-        }
-    }
-
-    private fun setupBottomNavigation() {
-        binding.bottomNavigationViewHome.setOnItemSelectedListener { menu ->
-            when(menu.itemId) {
-                R.id.home -> {
-                    HomeFragment().changeFragment()
-                }
-                R.id.community -> {
-//                    CommunityFragment().changeFragment()
-                }
-                R.id.my_page -> {
-//                    MyPageFragment().changeFragment()
-                }
-            }
-            return@setOnItemSelectedListener true
-        }
-    }
-
-    private fun Fragment.changeFragment() {
-        supportFragmentManager.beginTransaction().replace(R.id.fl_main,this).commit()
-    }
-}
